Create netlify _headers
Create a _headers file for Netlify to get Gatsby's service workers working.
The job of a service worker is to tell the browser to cache content, but if the browser caches the service worker file, you won’t see changes to the service worker itself, so you end up with stale content. A general best practice is to add a Cache-Control header when fetching the service worker file, explicitly telling the browser not to cache it.